+++ CHANGELOG.md
@@ -17,7 +17,9 @@ We use the following categories for changes:
 ## [Unreleased]
 
 ### Added
+
 - added tags to report [#sc-6938]
+- partial success flow [#sc-7483]
 
 ### Changed