+++ CHANGELOG.md
@@ -18,12 +18,28 @@ We use the following categories for changes:
 
 ### Added
 
+### Changed
+
+### Fixed
+
+### Removed
+
+### Deprecated
+
+### Security
+
+## [0.8.2] 2023-02-27
+
+### Added
+
 - added tags to report [#sc-6938]
 - partial success flow [#sc-7483]
 
 ### Changed
 
 ### Fixed
+
+- suppress analytics logger [#sc-7482]
 
 ### Removed